+++ CHANGELOG.md
@@ -6,6 +6,7 @@ All notable changes to this project will be documented in this file.
 
 ### Added
 - Added exception details for telemetry for SendMessage
+- Expose `OriginalMessageId` to `ChatSDK.onNewMessage()` to handle message ordering
 
 ### Changed
 - Uptake [@microsoft/ocsdk@0.5.13](https://www.npmjs.com/package/@microsoft/ocsdk/v/0.5.13)